requirements.txt
requests==2.28.2
python-dotenv==0.21.1
pandas==1.5.3
terminaltables==3.1.10
